Improve Paymill library handling
We've been using a tagged release of the Paymill PHP library for
a while, because we can't track the master branch anymore, since a new
major version had been merged.
The version parsing has never really been accurate, because the value we
parsed doesn't seem to be the actual version of the library. Now that we
know the exact version, let's introduce a constant and use that wherever
we can.